backend/requirements.txt
streamlit>=1.20.0
pymongo>=4.0.0
python-dotenv
pillow
exifread
opencv-python-headless
numpy
folium
streamlit-folium
requests
passlib[bcrypt]
geopy
